chore: organize codebase
