+++ src/lib/components/Homepage/hero.svelte
