+++ src/lib/components/Homepage/project_card.svelte
