+++ src/lib/components/common/icons/go.svelte
