+++ src/lib/components/common/icons/html.svelte
