+++ src/lib/components/common/icons/index.ts
@@ -1,0 +1,10 @@
+export { default as Html } from "./html.svelte"
+export { default as Css } from "./css.svelte"
+export { default as Sass } from "./sass.svelte"
+export { default as Javascript } from "./javascript.svelte"
+export { default as NodeJs } from "./node_js.svelte"
+export { default as ExpressJs } from "./express_js.svelte"
+export { default as Go } from "./go.svelte"
+export { default as Rust } from "./rust.svelte"
+export { default as Tailwind } from "./tailwind.svelte"
+export { default as Svelte } from "./svelte.svelte"
