+++ src/lib/components/common/icons/javascript.svelte
