+++ src/lib/components/common/icons/node_js.svelte
