+++ src/lib/components/common/icons/rust.svelte
