+++ src/lib/components/common/icons/sass.svelte
