+++ src/lib/components/common/icons/svelte.svelte
