+++ src/routes/+layout.svelte
@@ -1,9 +1,8 @@
 <script lang="ts">
+	import Navbar from '$lib/components/layouts/navbar.svelte';
+	import Footer from '$lib/components/layouts/footer.svelte';
 	import '../app.css';
 	import "@fontsource/poppins"
-
-	import Navbar from '$lib/components/navbar.svelte';
-	import Footer from '$lib/components/footer.svelte';
 </script>
 
 <Navbar />
